Implement SEO improvements with Next.js 15 and modern metadata API
