+++ app/page.tsx
@@ -1,23 +1,48 @@
-import { getAllPosts } from '@/app/lib/server-utils';
-import { generateHomeMetadata, generateHomeStructuredData } from '@/app/lib/seo';
-import Navigation from '@/app/components/Navigation';
-import NowPlaying from '@/app/components/NowPlaying';
-import ClientHomePage from '@/app/components/ClientHomePage';
+import { getPosts } from './lib/data';
+import { createMetadata } from './lib/metadata';
 import { Metadata } from 'next';
+import HomeClient from './components/HomeClient';
 
 export async function generateMetadata(): Promise<Metadata> {
-  const posts = await getAllPosts();
-  return generateHomeMetadata(posts);
+  return createMetadata({
+    title: 'pramit.gg',
+    description: 'A living, evolving journal of interests, projects, and experiences',
+    path: '',
+  });
 }
 
 export default async function Home() {
-  const posts = await getAllPosts();
+  const posts = await getPosts();
   
-  // Set first 3 posts as featured
-  const featuredPosts = posts.slice(0, 3);
-  const recentPosts = posts.slice(3);
-  
-  const structuredData = generateHomeStructuredData(posts);
+  const structuredData = {
+    '@context': 'https://schema.org',
+    '@type': 'Blog',
+    name: 'pramit.gg',
+    description: 'A living, evolving journal of interests, projects, and experiences',
+    url: 'https://pramit.gg',
+    author: {
+      '@type': 'Person',
+      name: 'Pramit Mazumder',
+      url: 'https://pramit.gg',
+    },
+    publisher: {
+      '@type': 'Organization',
+      name: 'pramit.gg',
+      url: 'https://pramit.gg',
+    },
+    blogPost: posts.slice(0, 5).map(post => ({
+      '@type': 'BlogPosting',
+      headline: post.title,
+      url: `https://pramit.gg/post/${post.slug}`,
+      datePublished: post.created_at,
+      dateModified: post.updated_at,
+      author: {
+        '@type': 'Person',
+        name: 'Pramit Mazumder',
+      },
+      keywords: post.tags.join(', '),
+    })),
+  };
 
   return (
     <>
@@ -27,14 +52,7 @@ export default async function Home() {
           __html: JSON.stringify(structuredData),
         }}
       />
-      
-      <Navigation />
-      <NowPlaying />
-      
-      <ClientHomePage 
-        posts={recentPosts} 
-        featuredPosts={featuredPosts} 
-      />
+      <HomeClient posts={posts} />
     </>
   );
 }
