+++ app/post/[id]/page.tsx
@@ -1,25 +1,21 @@
 import { notFound } from 'next/navigation';
-import { getPostBySlug, getAllPostSlugs } from '@/app/lib/server-utils';
-import { generatePostMetadata, generateStructuredData } from '@/app/lib/seo';
-import PostContent from './PostContent';
-import ClientViewTracker from '@/app/components/ClientViewTracker';
+import { getPost, getPostSlugs } from '@/app/lib/data';
+import { createPostMetadata, createStructuredData } from '@/app/lib/metadata';
 import { Metadata } from 'next';
+import PostClient from './PostClient';
 
 interface PostPageProps {
-  params: {
-    id: string;
-  };
+  params: Promise<{ id: string }>;
 }
 
 export async function generateStaticParams() {
-  const slugs = await getAllPostSlugs();
-  return slugs.map((slug) => ({
-    id: slug,
-  }));
+  const slugs = await getPostSlugs();
+  return slugs.map((slug) => ({ id: slug }));
 }
 
 export async function generateMetadata({ params }: PostPageProps): Promise<Metadata> {
-  const post = await getPostBySlug(params.id);
+  const { id } = await params;
+  const post = await getPost(id);
   
   if (!post) {
     return {
@@ -27,17 +23,18 @@ export async function generateMetadata({ params }: PostPageProps): Promise<Metad
     };
   }
 
-  return generatePostMetadata(post);
+  return createPostMetadata(post);
 }
 
 export default async function PostPage({ params }: PostPageProps) {
-  const post = await getPostBySlug(params.id);
+  const { id } = await params;
+  const post = await getPost(id);
 
   if (!post) {
     notFound();
   }
 
-  const structuredData = generateStructuredData(post);
+  const structuredData = createStructuredData(post);
 
   return (
     <>
@@ -47,16 +44,7 @@ export default async function PostPage({ params }: PostPageProps) {
           __html: JSON.stringify(structuredData),
         }}
       />
-      
-      <div className="min-h-screen bg-gradient-to-br from-void-black via-charcoal-black to-void-black">
-        <div className="absolute inset-0 bg-[radial-gradient(circle_at_20%_30%,rgba(120,119,198,0.03),transparent_50%)]" />
-        <div className="absolute inset-0 bg-[radial-gradient(circle_at_80%_70%,rgba(255,107,61,0.02),transparent_50%)]" />
-        
-        <div className="relative z-10 max-w-4xl mx-auto px-4 py-8">
-          <ClientViewTracker postId={post.id} />
-          <PostContent post={post} />
-        </div>
-      </div>
+      <PostClient post={post} />
     </>
   );
 }
