+++ app/sitemap.ts
@@ -1,11 +1,11 @@
-import { getAllPosts } from './lib/server-utils';
 import { MetadataRoute } from 'next';
+import { getPosts } from './lib/data';
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
-  const posts = await getAllPosts();
+  const posts = await getPosts();
   
-  // Static pages
-  const staticPages = [
+  // Static routes
+  const routes = [
     {
       url: 'https://pramit.gg',
       lastModified: new Date(),
@@ -26,13 +26,13 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
     },
   ];
 
-  // Dynamic post pages
-  const postPages = posts.map((post) => ({
+  // Dynamic post routes
+  const postRoutes = posts.map((post) => ({
     url: `https://pramit.gg/post/${post.slug}`,
     lastModified: new Date(post.updated_at),
     changeFrequency: 'weekly' as const,
     priority: 0.9,
   }));
 
-  return [...staticPages, ...postPages];
+  return [...routes, ...postRoutes];
 }
