+++ package.json
@@ -16,11 +16,11 @@
     "@uiw/react-md-editor": "^4.0.3",
     "date-fns": "^3.0.6",
     "dotenv": "^16.4.7",
-    "framer-motion": "^10.16.16",
-    "next": "14.0.4",
+    "framer-motion": "^12.23.5",
+    "next": "15.4.1",
     "qrcode": "^1.5.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-intersection-observer": "^9.5.3",
     "react-markdown": "^9.0.1",
     "react-masonry-css": "^1.0.16",
@@ -33,11 +33,11 @@
   },
   "devDependencies": {
     "@types/node": "^20.10.5",
-    "@types/react": "^18.2.45",
-    "@types/react-dom": "^18.2.18",
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.0",
     "autoprefixer": "^10.4.16",
     "eslint": "^8.56.0",
-    "eslint-config-next": "14.0.4",
+    "eslint-config-next": "15.4.1",
     "postcss": "^8.4.32",
     "tailwindcss": "^3.4.0",
     "ts-node": "^10.9.2",
